Add files via upload
Setup:
1. setup stepper motor normally
2. setup esp + oled screen normally
3. connect ground of esp to ground of arduino
4. connect tx of esp to rx of arduino
5. connect rx of esp to tx of arduino

this was an attempt to integrate between the stepper motor and esp. 
why would we need to do that?
because of time!
esp knows the time and now we need it to communicate to arduino at a particular time to start running the motor!

something in the communication is failing.
possible leads:
its possible that all this code needs to be integrated into a single file and uploaded all at once.
useful chat: https://chatgpt.com/share/a18e1c8a-8987-48d3-8337-17a559a9668d

uploading pic to drive called "integration motor and esp"

motor_continuous is the file that works on its own - turns the motor on, spinning it in same direction continously.